+++ zng-view/src/surface.rs
@@ -404,6 +404,7 @@ impl Surface {
                 FrameCapture::Mask(m) => Some(Some(m)),
             };
             if let Some(mask) = capture {
+                self.context.make_current();
                 captured_data = Some(images.frame_image_data(
                     renderer,
                     PxRect::from_size(self.size.to_px(self.scale_factor)),
@@ -417,6 +418,7 @@ impl Surface {
     }
 
     pub fn frame_image(&mut self, images: &mut ImageCache, mask: Option<ImageMaskMode>) -> ImageId {
+        self.context.make_current();
         images.frame_image(
             self.renderer.as_mut().unwrap(),
             PxRect::from_size(self.size.to_px(self.scale_factor)),
@@ -430,6 +432,7 @@ impl Surface {
 
     pub fn frame_image_rect(&mut self, images: &mut ImageCache, rect: PxRect, mask: Option<ImageMaskMode>) -> ImageId {
         let rect = PxRect::from_size(self.size.to_px(self.scale_factor)).intersection(&rect).unwrap();
+        self.context.make_current();
         images.frame_image(
             self.renderer.as_mut().unwrap(),
             rect,